Add files via upload
Additional modifications to C2SIM Core ontology addressing issues 3, 6, 24, and 36, plus resolution of issue 43: added datatype simulationRealtimeMultiple, defined as positive xsd:decimal; modified range of hasSimulationRealtimeMultiple and SystemCommandBody subclasses using that property to reference the new simulationRealtimeMultiple datatype; modified annotation comments on SystemCommandBody subclasses using the simulationRealtimeMultiple value to indicate it is decimal rather than integer; modified annotations on classes relating to record and playback regarding the use of hasName identifying the file of recorded messages; added to annotation on PlaybackStatusReport to describe the playback position (keeping the datatype generic for now); added to annotations on SystemCommandBody subclasses StartPlayback and StartRecording as being under control of the C2SIM server; added data property hasTasker as subproperty of hasActorReference with range UUIDBase; added data property hasTaskID as subproperty of hasTaskReference with range UUIDBase; for issue#6, added hasTasker and hasTaskID property axioms to SystemCommand subclass MagicMove and updated the annotation for MagicMove; for issue#43, created new SystemCommand subclasses CheckpointSave and CheckpointRestore, each with property axiom for hasName. Changed the class name of SystemCommandBody to SystemMessageBody and modified the annotation accordingly. Added a hasTimeSincePlaybackStart data property with value isoTimeDurationBase as a property axiom on PlaybackStatusReport (the report can provide a PlaybackPosition and/or this TimeSincePlaybackStart).